Boot from volume without image supplied
This patch allows for booting instances without supplying an image if
there is block device mapping supplied. It makes changes to nova api and
compute services to handle requests that do not have any image supplied.

Also it makes rescue and rebuild work with instances started from
volume.

Finally the patch introduces tests to make sure the system acts as
expected, and in the process fixes and refactors some old tests to make
them test for cases this new functionality can introduce.

This patch is intended to be a proof of concept and a first step towards
a more cleaner interface for booting from volumes, outlined in
https://etherpad.openstack.org/grizzly-boot-from-volumes.

Direct next step could be to introduce a 'volume' parameter for starting
instances.

This patch is coupled with I5ba9b0f35a5084aa91eca260f46cac83b8b6591e
that provides changes to the nova client

Change-Id: I530760cfaa5eb0cae590c7383e0840c6b3f896b9