Directly copy a file URL from glance.
That patch allows nova to use shutil to copy a file directly from the
glance store.  For this to work glance needs to enable direct_url
descriptions in the image metadata.  This is done by setting
show_image_direct_url to True in glance-api.conf.  The path used to back
glance also must match the path used to back nova instance storage.  This
should be a large performance increase for the cases where both glance
and nova instance storage are backed by a shared file system like gluster
or ceph.

A new option has been added to nova: allowed_direct_url_schemes
This option is set to a list of schemes that can be directly accessed.
For this first simple patch only file:// is valid, and the
implementation is simple and inline.  In the future it would be better
to make this into propagation modules which are dynamically loaded
according to the scheme.

DocImpact
Implements: blueprint direct-file-copy
Change-Id: I041b5524f8a0ea59e89ece0202f0503abf2f9d8e