Introduce the BlockDeviceDict class
This patch introduces the BlockDeviceDict class that will serve as
a wrapper around the dictionaries representing block devices.

The idea behind this change is to make it easier to transition to the
new database table layout without having to change all the code that
touches block device mappings.

blueprint: improve-block-device-handling

Change-Id: Iecdc6d6a17f50e7395d99204f753e2b3d7788a8d