BDM class and transformation functions
This patch adds the utility functions that will help us switch
between versions od block_device_mapping data.

In order to acomplish this, and to have a more structured aproach to
dealing with block device mapping data in the future, this patch
introduces a BlockDeviceDict class which is a thin wrapper around a
standard dict that makes it easier to reason about which fields are
present in the dictionary, and introduces said conversion routines.

Change-Id: I9370333059b8c9aaf92010470b8475a913d329b2