Allow block devices without device_name
This patch removes the check in the API that prevented having block
devices without 'device_name' and makes the compute manager capable of
guessing those names.

Before creating the instance, the manager will check each one of the
block devices that should be attached and verify that all of them have a
device_name assigned. If any of them doesn't, it will try to guess a
valid name using the method `get_next_device_name', which should be
implemented by the drivers, or by calling the generic method from the
utils module `get_device_name_for_instance'.

It also takes care of assigning the correct device name to the root
device (boot_index=0) to make it agree with the field `root_device_name'
kept in the instance.

Part of blueprint: improve-block-device-handling

Change-Id: I84541f8ff6e1b5978734e5def69946d014c66fdf