Add method for getting the CPU pinning constraint
We add a method that when passed flavor with extra spec containing
hw:cpu_policy key, and an optional NUMA topology instance,
calculates the CPU pinning constraint for the instance (or None if the
instance is not be pinned). In case admin doesn't specify a shared pinning
policy on the flavor, users are free to override it with image metadata
hw_cpu_policy, otherwise, attempts to override it will result in an
error.

This will be done at boot time and then persisted to be used later in
scheduling.

Change-Id: I1793ce0288aa97d7a7cac3df1ca84c48935ec46b
Blueprint: virt-driver-cpu-pinning