libvirt: Add method for getting host cpu_pinning
We can get this information from the libvirt capabilities XML, so we
add a method that returns it in the form of
virt.hardware.VirtCPUPinningTopology objects. We will exclude any CPUs not
found in the CONF.vcpu_pin_set.

We also fix how siblings were reported in the host capabilities fixture
in driver tests, as we now actually use this data.

Change-Id: I13388eda2a79d960f6e91377e32c9bedb490bdb1
Blueprint: virt-driver-cpu-pinning