Expose numa_topology to the resource tracker
This patch makes all the current drivers that implement
get_available_resource() return an additional 'cpu_pinning' field in
the resources dict. The libvirt driver will actually return useful data,
while other drivers only return 'None' at this time.

We also make the resource tracker check for the presence of this field.
The intent is that it will  be mandatory for all drivers to expose.

Change-Id: Ieee12706a1dddade610a2af2167c3f59838d9102
Blueprint: virt-driver-cpu-pinning