Hook CPU pinning checking into claims
This patch adds the ability to test claims against CPU pinning data exposed
by the host and instance. This is currently a no-op as no code in the
API will actually set any pinning constraints for the instance.
Since getting the CPU pinning information  of the instance may require an
additional db roundtrip, we try to avoid getting it unless absolutely
necessary.

The resource tracker will not do any actual usage tracking yet - this
will be done in a follow-up patch.

Change-Id: I2ee20843ce6b5dbade9b45bd711dd8e544a59433
Blueprint: virt-driver-cpu-pinning