API boot process sets CPU pinning for instances
This patch makes the API instance boot path look at the bits that can
contain requested CPU pinning, generate the desired topology and make sure
it is saved on the instance itself.

We also need to account for the flavor passed to
nova.compute.api.API.create not being an actual Flavor object, and
convert it to one.

Change-Id: I2d1866eb7c7d82e60a0757048db5ced60abcfe06
Blueprint: virt-driver-cpu-pinning