libvirt: Consider CPU pinning when booting
Consider CPU pinning request when booting an instance and make sure that
the proper XML is generated. If there was some pinning information
generated as part of the proposed NUMA configuration, this will make it
more specific as it will pin to the exact pCPU core as opposed to the
set that belongs to a host's NUMA node.

We also check any CPU topology that may have been exposed by the
cpu_pinning constraints, and make sure we use that instead of generating
one.

Change-Id: I717812c9f6f6bfab70317741b82d4c3ad47e26f3
Blueprint: virt-driver-cpu-pinning